ci/Jenkinsfile: switch to currentBuild.getBuildCauses()
access to currentBuild.rawBuild is not save and causes potential bugs
together with sh() calls. Besides that, access to rawBuild has to be
explicitely whitelisted for script security.

As a side effect, @NonCPS is not longer needed for handleBuildCause(),
so this should be restartable now.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>
(cherry picked from commit 413701f85362855ba41186dbc692c32f2c533b85)