ci/Jenkinsfile: switch to currentBuild.getBuildCauses()
access to currentBuild.rawBuild is not save and causes potential bugs
together with sh() calls. Besides that, access to rawBuild has to be
explicitely whitelisted for script security.

As a side effect, @NonCPS is not longer needed for handleBuildCause()
and handleGlobalData, so this should be restartable now.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>
(cherry picked from commit 2a975cd971921c33caf05f1c3b4cce5b68e72bb5)