ci/Jenkinsfile: switch to shared global SSTATE_DIR
When sharing SSTATE across the node we can save a bunch
of (re)build time. We expect SSTATE_DIR to be configured in the
node specific site.conf.

The downside ist, that we should not unconditionally delete
files in SSTATE_CACHE. So disable nightly cache cleanup for now
an give a warning, if the CLEAN_CACHES param is set.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>