ci/Jenkinsfile: use long options for `git branch`
The meaning of the `-l` option of `git branch` has changed between Git 2.17
and 2.25. The usage as found in our Jenkinsfile only works with the newer
version, breaking builds on Ubuntu 18.04 and older.

Use the long option `--list` (and also change `-r` to `--remotes`), which
is portable between Git versions.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>