ci: add tqma64xxl support
Add tqma64 to the list of TI machines. The tqma65_machines variable is
leftover from a previous commit; clean it up.

This will build tq-image-weston with the dumpling-wayland-ti DISTRO.
While adding Wayland to the TQMa64xxL builds is not very useful (as it
doesn't have any display output), it still makes sense to reuse as much of
the TQMa65xx build as possible instead of adding another config variant.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>