ci/create_update_commit.py: Add regular git repository support
This mode is intended to be used for updating individual packages, e.g.
linux-tq-5.15. Different to submodule mode any intended changes for the
commit have to be added to the index already.
The package name and the commit range for the git repository have to be
passed as well.
Note: If the range includes merges, only commits from the target branch
are included by default, see '--first-parent' for git-log.

Signed-off-by: Alexander Stein <alexander.stein@ew.tq-group.com>