ci/git-archive-all.sh: improve script
After commit 49d879a5 ("ci/git-archive-all: Fail if a command fails")
the script will abort on unhandled error. To prevent this to
happen without a notification add an ERR trap handler with
output of $LINENO for error.

Provide an error reporting function to enable error messages to
stderr. Additionally some errors are handled to give better error
messages on abort.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>