ci/build_all: handle deploy for linux-dummy
We depend on generic artifact names for rootfs as *.tgz,
kernel image and device tree names for sending artifacts
to automated testing. On the other hand there are valid
use cases where to build with linux-dummy.

Do not fail in case no device tree and kernel images can be found.
The build should have succeeded if we reached deploy stage. The
test step shall fail in case of misconfiguration.

To prevent multiple checks rely on find for copy. Since the
artifact names without annotation are links, restrict the
find to links.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>