ci/create_update_commit.py: Allow unsetting --logargs default
By default --logargs is set to '--first-parent', this only shows direct
commits on the main branch but hides commits integrated using a merge.
While the default is good for huge repositories like Linux kernel,
this hides commits in small repositories where code is only merged using
Merge Requests on Gitlab as only the merge commits are listed in history.
Allow unsetting the default by specifying --logargs=""
As an empty logargs argument to git module is invalid, is has to be set to
None explicitly.

Signed-off-by: Alexander Stein <alexander.stein@ew.tq-group.com>