ci: git-revision-name.sh: allow git-describe and git show-ref to fail
The output of these commands is checked later, and error messages are
suppressed, so it doesn't make sense to have the whole script fail due
to `set -e` on errors (which will happen in a repo without any tags).

For `git rev-parse --verify HEAD`, the redirect of error messages is
removed instead - there should always be a HEAD, so there shouldn't be
any error messages to suppress.

The script will still fail as intended if the passed ref is invalid or
doesn't point at the current HEAD, but using the script in a new repo
that doesn't have any tags yet is possible again.

Fixes: df47540de28e ("ci/git-revision-name.sh: rewrite to verify a given revision")
Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>